Move some utils over to actions
