+++ packages/desktop/components/popups/ImportErc20TokenFormPopup.svelte
@@ -4,6 +4,7 @@
     import { localize } from '@core/i18n'
     import { ERC20_TOKEN_ADDRESS_LENGTH } from '@core/layer-2'
     import { getErc20TokenMetadata } from '@core/layer-2/utils'
+    import { network } from '@core/network'
     import { HEXADECIMAL_PREFIX, HEXADECIMAL_REGEXP } from '@core/utils'
 
     import { closePopup } from '@desktop/auxiliary/popup'
@@ -27,7 +28,7 @@
 
         if (validate()) {
             try {
-                const erc20TokenMetadata = await getErc20TokenMetadata(tokenAddress, chainId)
+                const erc20TokenMetadata = await getErc20TokenMetadata(tokenAddress, chainId, $network)
                 if (erc20TokenMetadata) {
                     addNewTrackedTokenToActiveProfile(chainId, tokenAddress, erc20TokenMetadata)
                     showNotification({
