+++ packages/shared/src/lib/core/account/actions/index.ts
@@ -1,6 +1,7 @@
 export * from './buildAccountState'
 export * from './buildAccountStateAndPersistedData'
 export * from './createNewAccount'
+export * from './isAccountAddress'
 export * from './loadAccount'
 export * from './resetSelectedAccountIndex'
 export * from './setNextSelectedAccount'
@@ -8,3 +9,4 @@ export * from './setSelectedAccount'
 export * from './syncBalance'
 export * from './syncVotingPower'
 export * from './tryCreateAdditionalAccount'
+export * from './validateAccountName'
