+++ packages/shared/src/lib/core/account/utils/index.ts
@@ -9,4 +9,3 @@ export * from './sumBalanceForAccounts'
 export * from './sumTotalFromOutputs'
 export * from './syncAccountsInParallel'
 export * from './syncAccountsInSeries'
-export * from './validateAccountName'
