+++ packages/shared/src/lib/core/layer-2/utils/getErc20TokenMetadata.ts
@@ -1,13 +1,14 @@
-import { network } from '@core/network/stores'
-import { IErc20Metadata, TokenStandard } from '@core/wallet'
-import { get } from 'svelte/store'
+import { INetwork } from '@core/network/interfaces'
+import { TokenStandard } from '@core/wallet/enums'
+import { IErc20Metadata } from '@core/wallet/interfaces'
 import { ContractType } from '../enums'
 
 export async function getErc20TokenMetadata(
     tokenAddress: string,
-    chainId: number
+    chainId: number,
+    network: INetwork
 ): Promise<IErc20Metadata | undefined> {
-    const chain = get(network)?.getChain(chainId)
+    const chain = network?.getChain(chainId)
     const contract = chain?.getContract(ContractType.Erc20, tokenAddress)
     if (contract) {
         const [name, symbol, decimals] = await Promise.all([
