+++ packages/shared/src/lib/core/wallet/actions/requestPersistedAsset.ts
@@ -1,4 +1,4 @@
-import { getErc20TokenMetadata } from '@core/layer-2'
+import { getErc20TokenMetadata } from '@core/layer-2/utils'
 import { validateEthereumAddress } from '@core/utils/crypto/utils'
 import { OFFICIAL_TOKEN_IDS } from '../constants'
 import { NotVerifiedStatus, VerifiedStatus } from '../enums'
@@ -6,6 +6,8 @@ import { buildPersistedAssetFromMetadata } from '../helpers'
 import { IErc20Metadata, IIrc30Metadata, IPersistedAsset } from '../interfaces'
 import { AssetVerification } from '../types'
 import { getIrc30MetadataFromFoundryOutput } from '../utils'
+import { activeAccounts } from '@core/profile/stores'
+import { get } from 'svelte/store'
 
 export async function requestPersistedAsset(tokenId: string, chainId?: number): Promise<IPersistedAsset | undefined> {
     let tokenMetadata: IIrc30Metadata | IErc20Metadata | undefined
@@ -17,7 +19,10 @@ export async function requestPersistedAsset(tokenId: string, chainId?: number): 
             // do nothing
         }
     } else {
-        tokenMetadata = await getIrc30MetadataFromFoundryOutput(tokenId)
+        const account = get(activeAccounts)?.[0]
+        if (account) {
+            tokenMetadata = await getIrc30MetadataFromFoundryOutput(tokenId, account)
+        }
     }
     if (tokenMetadata) {
         const verification: AssetVerification = OFFICIAL_TOKEN_IDS.includes(tokenId)
