+++ packages/shared/src/lib/core/wallet/utils/index.ts
@@ -18,7 +18,6 @@ export * from './getSenderAddressFromUnlockCondition'
 export * from './getSubjectFromAddress'
 export * from './getUnitFromTokenMetadata'
 export * from './hashOutputId'
-export * from './isAccountAddress'
 export * from './isReservedTagKeyword'
 export * from './isSubjectInternal'
 export * from './parseGovernanceMetadata'
